requirements.txt
requests>=2.26.0
beautifulsoup4>=4.10.0
python-dateutil>=2.8.2
